refactor(sdk): move auth-error from components to utils
It is a mapping from a connect error to copy, with no React in it, so
it sits with the other client utils and is imported by path like them.
