--- web/sdk/client/components/auth-error/index.ts
@@ -1,2 +1,0 @@
-export { describeAuthError } from './auth-error';
-export type { AuthRejection } from './auth-error';
