+++ web/sdk/client/utils/auth-error.ts
