+++ web/sdk/client/views/auth/magic-link/magic-link-view.tsx
@@ -12,7 +12,7 @@ import {
   type AuthContainerProps
 } from '~/client/components/auth-container';
 import { AuthHeader } from '~/client/components/auth-header';
-import { describeAuthError } from '~/client/components/auth-error';
+import { describeAuthError } from '~/client/utils/auth-error';
 import styles from './magic-link-view.module.css';
 
 export type MagicLinkViewProps = ComponentPropsWithRef<'div'> &
